requirement.txt
flask
flask-sqlalchemy
flask-migrate
joblib
pandas
numpy
xgboost
scikit-learn
pymysql
cryptography
argon2-cffi 
PyJWT
dotenv